// client/app/models/user/user.model.ts
export class User {
    id: String;
    firstName: String;
    lastName: String;
    provider: String;
    providerId: String;
    providerData: Object;
    email: String;
    photo: String;
    username: String;
    password: String;
    role: String;
    created: Date;
    website: String;
    fullName: String;
}